Make GDPR banner work with IE11 as well
* Switch to using "position: fixed".
* Remove the translate which was making the horizontal scroll appear.

Since the banner was moved inside the page-wrapper, the translate was not really doing anything useful. It was initially
added to shift the page down 36px so that the banner does not cover the content. However, when moved inside the page-wrapper,
it does cover the content. So, translate is not necessary.

In this case, "position: fixed" works as well as "position: sticky" anyway. So, using the more standard one.